Move clowder yaml utilitiies into one file
