+++ clowder/clowder_controller.py
@@ -8,10 +8,10 @@ import signal
 import sys
 
 import psutil
-from clowder.yaml.parsing import parse_yaml
-from clowder.yaml.printing import print_yaml
-from clowder.yaml.saving import save_yaml
-from clowder.yaml.validation import (
+from clowder.clowder_yaml import (
+    parse_yaml,
+    print_yaml,
+    save_yaml,
     validate_yaml,
     validate_yaml_import
 )
@@ -22,7 +22,7 @@ import clowder.utility.formatting as fmt
 from clowder.error.clowder_error import ClowderError
 from clowder.group import Group
 from clowder.source import Source
-from clowder.yaml.load import (
+from clowder.clowder_yaml import (
     load_yaml_base,
     load_yaml_import
 )
