+++ clowder/clowder_yaml.py
@@ -1,8 +1,115 @@
-"""Clowder yaml validation"""
+"""clowder.yaml parsing and functionality"""
 
-import clowder.utility.formatting as fmt
+from __future__ import print_function
+import os
+import sys
+import yaml
+
+from termcolor import colored
+
 from clowder.error.clowder_error import ClowderError
-from clowder.yaml.parsing import parse_yaml
+import clowder.utility.formatting as fmt
+
+
+def load_yaml_base(parsed_yaml, combined_yaml):
+    """Load clowder from base yaml file"""
+    combined_yaml['defaults'] = parsed_yaml['defaults']
+    if 'depth' not in parsed_yaml['defaults']:
+        combined_yaml['defaults']['depth'] = 0
+    combined_yaml['sources'] = parsed_yaml['sources']
+    combined_yaml['groups'] = parsed_yaml['groups']
+
+
+def load_yaml_import(parsed_yaml, combined_yaml):
+    """Load clowder from import yaml file"""
+    if 'defaults' in parsed_yaml:
+        _load_yaml_import_defaults(parsed_yaml['defaults'], combined_yaml['defaults'])
+    if 'sources' in parsed_yaml:
+        _load_yaml_import_sources(parsed_yaml['sources'], combined_yaml['sources'])
+    if 'groups' in parsed_yaml:
+        _load_yaml_import_groups(parsed_yaml['groups'], combined_yaml['groups'])
+
+
+def parse_yaml(yaml_file):
+    """Parse yaml file"""
+    if os.path.isfile(yaml_file):
+        try:
+            with open(yaml_file) as raw_file:
+                parsed_yaml = yaml.safe_load(raw_file)
+                if parsed_yaml is None:
+                    print(fmt.invalid_yaml_error())
+                    print(fmt.empty_yaml_error(yaml_file) + '\n')
+                    sys.exit(1)
+                return parsed_yaml
+        except yaml.YAMLError:
+            fmt.open_file_error(yaml_file)
+            sys.exit(1)
+        except (KeyboardInterrupt, SystemExit):
+            sys.exit(1)
+    else:
+        print()
+        print(fmt.missing_yaml_error())
+        print()
+        sys.exit(1)
+
+
+def print_yaml(root_directory):
+    """Print current clowder yaml"""
+    yaml_file = os.path.join(root_directory, 'clowder.yaml')
+    parsed_yaml = parse_yaml(yaml_file)
+    yaml_files = []
+    while True:
+        yaml_files.append(yaml_file)
+        if 'import' not in parsed_yaml:
+            break
+        imported_yaml = parsed_yaml['import']
+        if imported_yaml == 'default':
+            yaml_file = os.path.join(root_directory, '.clowder', 'clowder.yaml')
+        else:
+            yaml_file = os.path.join(root_directory, '.clowder', 'versions', imported_yaml, 'clowder.yaml')
+        parsed_yaml = parse_yaml(yaml_file)
+    for yaml_file in yaml_files:
+        if os.path.isfile(yaml_file):
+            try:
+                with open(yaml_file) as raw_file:
+                    contents = raw_file.read()
+                    print('-' * 80)
+                    if os.path.islink(yaml_file):
+                        path = fmt.symlink_target(yaml_file)
+                        path = fmt.remove_prefix(path, root_directory)
+                        path = fmt.remove_prefix(path, '/')
+                        print()
+                        print(fmt.path('clowder.yaml') + ' -> ' + fmt.path(path))
+                        print()
+                    else:
+                        path = fmt.remove_prefix(yaml_file, root_directory)
+                        path = fmt.remove_prefix(path, '/')
+                        print('\n' + fmt.path(path) + '\n')
+                    print(contents)
+            except IOError as err:
+                fmt.open_file_error(yaml_file)
+                print(err)
+                sys.exit(1)
+            except (KeyboardInterrupt, SystemExit):
+                sys.exit(1)
+
+
+def save_yaml(yaml_output, yaml_file):
+    """Save yaml file to disk"""
+    if not os.path.isfile(yaml_file):
+        try:
+            with open(yaml_file, 'w') as raw_file:
+                print(" - Save yaml to file")
+                yaml.safe_dump(yaml_output, raw_file, default_flow_style=False, indent=4)
+        except yaml.YAMLError:
+            fmt.save_file_error(yaml_file)
+            sys.exit(1)
+        except (KeyboardInterrupt, SystemExit):
+            sys.exit(1)
+    else:
+        fmt.file_exists_error(yaml_file)
+        print()
+        sys.exit(1)
 
 
 def validate_yaml(yaml_file):
@@ -16,19 +123,19 @@ def validate_yaml(yaml_file):
     if 'defaults' not in parsed_yaml:
         error = fmt.missing_entry_error('defaults', fmt.yaml_file('clowder.yaml'), yaml_file)
         raise ClowderError(error)
-    validate_yaml_defaults(parsed_yaml['defaults'], yaml_file)
+    _validate_yaml_defaults(parsed_yaml['defaults'], yaml_file)
     del parsed_yaml['defaults']
 
     if 'sources' not in parsed_yaml:
         error = fmt.missing_entry_error('sources', fmt.yaml_file('clowder.yaml'), yaml_file)
         raise ClowderError(error)
-    validate_yaml_sources(parsed_yaml['sources'], yaml_file)
+    _validate_yaml_sources(parsed_yaml['sources'], yaml_file)
     del parsed_yaml['sources']
 
     if 'groups' not in parsed_yaml:
         error = fmt.missing_entry_error('groups', fmt.yaml_file('clowder.yaml'), yaml_file)
         raise ClowderError(error)
-    validate_yaml_groups(parsed_yaml['groups'], yaml_file)
+    _validate_yaml_groups(parsed_yaml['groups'], yaml_file)
     del parsed_yaml['groups']
 
     if parsed_yaml:
@@ -51,15 +158,15 @@ def validate_yaml_import(yaml_file):
         raise ClowderError(error)
 
     if 'defaults' in parsed_yaml:
-        validate_yaml_import_defaults(parsed_yaml['defaults'], yaml_file)
+        _validate_yaml_import_defaults(parsed_yaml['defaults'], yaml_file)
         del parsed_yaml['defaults']
 
     if 'sources' in parsed_yaml:
-        validate_yaml_sources(parsed_yaml['sources'], yaml_file)
+        _validate_yaml_sources(parsed_yaml['sources'], yaml_file)
         del parsed_yaml['sources']
 
     if 'groups' in parsed_yaml:
-        validate_yaml_import_groups(parsed_yaml['groups'], yaml_file)
+        _validate_yaml_import_groups(parsed_yaml['groups'], yaml_file)
         del parsed_yaml['groups']
 
     if parsed_yaml:
@@ -67,347 +174,96 @@ def validate_yaml_import(yaml_file):
         raise ClowderError(error)
 
 
-def validate_yaml_import_defaults(defaults, yaml_file):
-    """Validate clowder.yaml defaults with an import"""
-    _validate_type_dict(defaults, 'defaults', yaml_file)
-    if 'recursive' in defaults:
-        _validate_type_bool(defaults['recursive'], 'recursive', yaml_file)
-        del defaults['recursive']
-    if 'ref' in defaults:
-        _validate_type_str(defaults['ref'], 'ref', yaml_file)
-        if not _valid_ref_type(defaults['ref']):
-            error = fmt.invalid_ref_error(defaults['ref'], yaml_file)
-            raise ClowderError(error)
-        del defaults['ref']
-    if 'remote' in defaults:
-        _validate_type_str(defaults['remote'], 'remote', yaml_file)
-        del defaults['remote']
-    if 'source' in defaults:
-        _validate_type_str(defaults['source'], 'source', yaml_file)
-        del defaults['source']
-    if 'depth' in defaults:
-        _validate_type_depth(defaults['depth'], yaml_file)
-        del defaults['depth']
-    if defaults:
-        error = fmt.invalid_entries_error('defaults', defaults, yaml_file)
-        raise ClowderError(error)
-
-
-def validate_yaml_defaults(defaults, yaml_file):
-    """Validate defaults in clowder loaded from yaml file"""
-    _validate_type_dict(defaults, 'defaults', yaml_file)
-    if not defaults:
-        error = fmt.invalid_entries_error('defaults', defaults, yaml_file)
-        raise ClowderError(error)
-
-    if 'ref' not in defaults:
-        error = fmt.missing_entry_error('ref', 'defaults', yaml_file)
-        raise ClowderError(error)
-    _validate_type_str(defaults['ref'], 'ref', yaml_file)
-    if not _valid_ref_type(defaults['ref']):
-        error = fmt.invalid_ref_error(defaults['ref'], yaml_file)
-        raise ClowderError(error)
-    del defaults['ref']
-
-    if 'remote' not in defaults:
-        error = fmt.missing_entry_error('remote', 'defaults', yaml_file)
-        raise ClowderError(error)
-    _validate_type_str(defaults['remote'], 'remote', yaml_file)
-    del defaults['remote']
-
-    if 'source' not in defaults:
-        error = fmt.missing_entry_error('source', 'defaults', yaml_file)
-        raise ClowderError(error)
-    _validate_type_str(defaults['source'], 'source', yaml_file)
-    del defaults['source']
-
-    validate_yaml_defaults_optional(defaults, yaml_file)
-
-    if defaults:
-        error = fmt.invalid_entries_error('defaults', defaults, yaml_file)
-        raise ClowderError(error)
-
-
-def validate_yaml_defaults_optional(defaults, yaml_file):
-    """Validate defaults optional args in clowder loaded from yaml file"""
-    if 'depth' in defaults:
-        _validate_type_depth(defaults['depth'], yaml_file)
-        del defaults['depth']
-
-    if 'recursive' in defaults:
-        _validate_type_bool(defaults['recursive'], 'recursive', yaml_file)
-        del defaults['recursive']
-
-
-def validate_yaml_fork(fork, yaml_file):
-    """Validate fork in clowder loaded from yaml file"""
-    _validate_type_dict(fork, 'fork', yaml_file)
-    if not fork:
-        error = fmt.invalid_entries_error('fork', fork, yaml_file)
-        raise ClowderError(error)
-
-    if 'name' not in fork:
-        error = fmt.missing_entry_error('name', 'fork', yaml_file)
-        raise ClowderError(error)
-    _validate_type_str(fork['name'], 'name', yaml_file)
-    del fork['name']
-
-    if 'remote' not in fork:
-        error = fmt.missing_entry_error('remote', 'fork', yaml_file)
-        raise ClowderError(error)
-    _validate_type_str(fork['remote'], 'remote', yaml_file)
-    del fork['remote']
-
-    if fork:
-        error = fmt.invalid_entries_error('fork', fork, yaml_file)
-        raise ClowderError(error)
-
-
-def validate_yaml_import_groups(groups, yaml_file):
-    """Validate groups in clowder loaded from yaml file with import"""
-    _validate_type_list(groups, 'groups', yaml_file)
-    if not groups:
-        error = fmt.invalid_entries_error('groups', groups, yaml_file)
-        raise ClowderError(error)
-
-    for group in groups:
-        validate_yaml_import_group(group, yaml_file)
-
-
-def validate_yaml_groups(groups, yaml_file):
-    """Validate groups in clowder loaded from yaml file"""
-    _validate_type_list(groups, 'groups', yaml_file)
-    if not groups:
-        error = fmt.invalid_entries_error('groups', groups, yaml_file)
-        raise ClowderError(error)
-
-    for group in groups:
-        validate_yaml_group(group, yaml_file)
-
-
-def validate_yaml_import_project(project, yaml_file):
-    """Validate project in clowder loaded from yaml file with import"""
-    _validate_type_dict(project, 'project', yaml_file)
-    if not project:
-        error = fmt.invalid_entries_error('project', project, yaml_file)
-        raise ClowderError(error)
-
-    if 'name' not in project:
-        error = fmt.missing_entry_error('name', 'project', yaml_file)
-        raise ClowderError(error)
-    _validate_type_str(project['name'], 'name', yaml_file)
-    del project['name']
-
-    if not project:
-        error = fmt.invalid_entries_error('project', project, yaml_file)
-        raise ClowderError(error)
-
-    if 'path' in project:
-        _validate_type_str(project['path'], 'path', yaml_file)
-        del project['path']
-
-    validate_yaml_project_optional(project, yaml_file)
-
-    if project:
-        error = fmt.invalid_entries_error('project', project, yaml_file)
-        raise ClowderError(error)
-
-
-def validate_yaml_import_group(group, yaml_file):
-    """Validate group in clowder loaded from yaml file with import"""
-    _validate_type_dict(group, 'group', yaml_file)
-    if not group:
-        error = fmt.invalid_entries_error('group', group, yaml_file)
-        raise ClowderError(error)
-
-    if 'name' not in group:
-        error = fmt.missing_entry_error('name', 'group', yaml_file)
-        raise ClowderError(error)
-    _validate_type_str(group['name'], 'name', yaml_file)
-    del group['name']
-
-    if not group:
-        error = fmt.invalid_entries_error('group', group, yaml_file)
-        raise ClowderError(error)
-
-    if 'projects' in group:
-        validate_yaml_projects(group['projects'], yaml_file, is_import=True)
-        del group['projects']
-
-    if 'recursive' in group:
-        _validate_type_bool(group['recursive'], 'recursive', yaml_file)
-        del group['recursive']
-
-    if 'ref' in group:
-        _validate_type_str(group['ref'], 'ref', yaml_file)
-        if not _valid_ref_type(group['ref']):
-            error = fmt.invalid_ref_error(group['ref'], yaml_file)
-            raise ClowderError(error)
-        del group['ref']
-
-    if 'remote' in group:
-        _validate_type_str(group['remote'], 'remote', yaml_file)
-        del group['remote']
-
-    if 'source' in group:
-        _validate_type_str(group['source'], 'source', yaml_file)
-        del group['source']
-
-    if 'depth' in group:
-        _validate_type_depth(group['depth'], yaml_file)
-        del group['depth']
-
-    if group:
-        error = fmt.invalid_entries_error('group', group, yaml_file)
-        raise ClowderError(error)
-
-
-def validate_yaml_group(group, yaml_file):
-    """Validate group in clowder loaded from yaml file"""
-    _validate_type_dict(group, 'group', yaml_file)
-    if not group:
-        error = fmt.invalid_entries_error('group', group, yaml_file)
-        raise ClowderError(error)
-
-    if 'name' not in group:
-        error = fmt.missing_entry_error('name', 'group', yaml_file)
-        raise ClowderError(error)
-    _validate_type_str(group['name'], 'name', yaml_file)
-    del group['name']
-
-    if 'projects' not in group:
-        error = fmt.missing_entry_error('projects', 'group', yaml_file)
-        raise ClowderError(error)
-    validate_yaml_projects(group['projects'], yaml_file, is_import=False)
-    del group['projects']
-
-    if 'recursive' in group:
-        _validate_type_bool(group['recursive'], 'recursive', yaml_file)
-        del group['recursive']
-
-    if 'ref' in group:
-        _validate_type_str(group['ref'], 'ref', yaml_file)
-        if not _valid_ref_type(group['ref']):
-            error = fmt.invalid_ref_error(group['ref'], yaml_file)
-            raise ClowderError(error)
-        del group['ref']
-
-    if 'remote' in group:
-        _validate_type_str(group['remote'], 'remote', yaml_file)
-        del group['remote']
-
-    if 'source' in group:
-        _validate_type_str(group['source'], 'source', yaml_file)
-        del group['source']
-
-    if 'depth' in group:
-        _validate_type_depth(group['depth'], yaml_file)
-        del group['depth']
-
-    if group:
-        error = fmt.invalid_entries_error('group', group, yaml_file)
-        raise ClowderError(error)
-
-
-def validate_yaml_project(project, yaml_file):
-    """Validate project in clowder loaded from yaml file"""
-    _validate_type_dict(project, 'project', yaml_file)
-    if not project:
-        error = fmt.invalid_entries_error('project', project, yaml_file)
-        raise ClowderError(error)
-
-    if 'name' not in project:
-        error = fmt.missing_entry_error('name', 'project', yaml_file)
-        raise ClowderError(error)
-    _validate_type_str(project['name'], 'name', yaml_file)
-    del project['name']
-
-    if 'path' not in project:
-        error = fmt.missing_entry_error('path', 'project', yaml_file)
-        raise ClowderError(error)
-    _validate_type_str(project['path'], 'path', yaml_file)
-    del project['path']
-
-    validate_yaml_project_optional(project, yaml_file)
-
-    if project:
-        error = fmt.invalid_entries_error('project', project, yaml_file)
-        raise ClowderError(error)
-
-
-def validate_yaml_project_optional(project, yaml_file):
-    """Validate optional args in project in clowder loaded from yaml file"""
-    if 'remote' in project:
-        _validate_type_str(project['remote'], 'remote', yaml_file)
-        del project['remote']
-
-    if 'recursive' in project:
-        _validate_type_bool(project['recursive'], 'recursive', yaml_file)
-        del project['recursive']
-
-    if 'ref' in project:
-        _validate_type_str(project['ref'], 'ref', yaml_file)
-        if not _valid_ref_type(project['ref']):
-            error = fmt.invalid_ref_error(project['ref'], yaml_file)
-            raise ClowderError(error)
-        del project['ref']
-
-    if 'source' in project:
-        _validate_type_str(project['source'], 'source', yaml_file)
-        del project['source']
-
-    if 'depth' in project:
-        _validate_type_depth(project['depth'], yaml_file)
-        del project['depth']
-
-    if 'fork' in project:
-        fork = project['fork']
-        validate_yaml_fork(fork, yaml_file)
-        del project['fork']
-
-
-def validate_yaml_projects(projects, yaml_file, is_import):
-    """Validate projects in clowder loaded from yaml file"""
-    _validate_type_list(projects, 'projects', yaml_file)
-    if not projects:
-        error = fmt.invalid_entries_error('projects', projects, yaml_file)
-        raise ClowderError(error)
-
-    for project in projects:
-        if is_import:
-            validate_yaml_import_project(project, yaml_file)
-        else:
-            validate_yaml_project(project, yaml_file)
-
-
-def validate_yaml_sources(sources, yaml_file):
-    """Validate sources in clowder loaded from yaml file"""
-    _validate_type_list(sources, 'sources', yaml_file)
-    if not sources:
-        error = fmt.invalid_entries_error('sources', sources, yaml_file)
-        raise ClowderError(error)
-
-    for source in sources:
-        _validate_type_dict(source, 'source', yaml_file)
-        if not source:
-            error = fmt.invalid_entries_error('source', source, yaml_file)
-            raise ClowderError(error)
-
-        if 'name' not in source:
-            error = fmt.missing_entry_error('name', 'source', yaml_file)
-            raise ClowderError(error)
-        _validate_type_str(source['name'], 'name', yaml_file)
-        del source['name']
-
-        if 'url' not in source:
-            error = fmt.missing_entry_error('url', 'source', yaml_file)
-            raise ClowderError(error)
-        _validate_type_str(source['url'], 'url', yaml_file)
-        del source['url']
-
-        if source:
-            error = fmt.invalid_entries_error('source', source, yaml_file)
-            raise ClowderError(error)
+def _load_yaml_import_defaults(imported_defaults, defaults):
+    """Load clowder projects from imported group"""
+    if 'recursive' in imported_defaults:
+        defaults['recursive'] = imported_defaults['recursive']
+    if 'ref' in imported_defaults:
+        defaults['ref'] = imported_defaults['ref']
+    if 'remote' in imported_defaults:
+        defaults['remote'] = imported_defaults['remote']
+    if 'source' in imported_defaults:
+        defaults['source'] = imported_defaults['source']
+    if 'depth' in imported_defaults:
+        defaults['depth'] = imported_defaults['depth']
+
+
+def _load_yaml_import_groups(imported_groups, groups):
+    """Load clowder groups from import yaml"""
+    group_names = [g['name'] for g in groups]
+    for imported_group in imported_groups:
+        if imported_group['name'] not in group_names:
+            groups.append(imported_group)
+            continue
+        combined_groups = []
+        for group in groups:
+            if group['name'] == imported_group['name']:
+                if 'recursive' in imported_groups:
+                    group['recursive'] = imported_groups['recursive']
+                if 'ref' in imported_group:
+                    group['ref'] = imported_group['ref']
+                if 'remote' in imported_group:
+                    group['remote'] = imported_group['remote']
+                if 'source' in imported_group:
+                    group['source'] = imported_group['source']
+                if 'depth' in imported_group:
+                    group['depth'] = imported_group['depth']
+                if 'projects' in imported_group:
+                    _load_yaml_import_projects(imported_group['projects'], group['projects'])
+            combined_groups.append(group)
+        groups = combined_groups
+
+
+def _load_yaml_import_projects(imported_projects, projects):
+    """Load clowder projects from imported group"""
+    project_names = [p['name'] for p in projects]
+    for imported_project in imported_projects:
+        if imported_project['name'] not in project_names:
+            if 'path' not in imported_project:
+                # error = fmt.invalid_entries_error('defaults', defaults, yaml_file)
+                error = colored(' - Missing path in new project', 'red')
+                print(fmt.invalid_yaml_error())
+                print(fmt.error(error))
+                sys.exit(1)
+            projects.append(imported_project)
+            continue
+        combined_projects = []
+        for project in projects:
+            if project['name'] != imported_project['name']:
+                combined_projects.append(project)
+                continue
+            if 'path' in imported_project:
+                project['path'] = imported_project['path']
+            if 'depth' in imported_project:
+                project['depth'] = imported_project['depth']
+            if 'recursive' in imported_project:
+                project['recursive'] = imported_project['recursive']
+            if 'ref' in imported_project:
+                project['ref'] = imported_project['ref']
+            if 'remote' in imported_project:
+                project['remote'] = imported_project['remote']
+            if 'fork' in imported_project:
+                project['fork'] = imported_project['fork']
+            if 'source' in imported_project:
+                project['source'] = imported_project['source']['name']
+            combined_projects.append(imported_project)
+        projects = combined_projects
+
+
+def _load_yaml_import_sources(imported_sources, sources):
+    """Load clowder sources from import yaml"""
+    source_names = [s['name'] for s in sources]
+    for imported_source in imported_sources:
+        if imported_source['name'] not in source_names:
+            sources.append(imported_source)
+            continue
+        combined_sources = []
+        for source in sources:
+            if source.name == imported_source['name']:
+                combined_sources.append(imported_source)
+            else:
+                combined_sources.append(source)
+        sources = combined_sources
 
 
 def _valid_ref_type(ref):
@@ -458,3 +314,346 @@ def _validate_type_str(value, name, yaml_file):
     if not isinstance(value, str):
         error = fmt.not_string_error(name, yaml_file)
         raise ClowderError(error)
+
+
+def _validate_yaml_import_defaults(defaults, yaml_file):
+    """Validate clowder.yaml defaults with an import"""
+    _validate_type_dict(defaults, 'defaults', yaml_file)
+    if 'recursive' in defaults:
+        _validate_type_bool(defaults['recursive'], 'recursive', yaml_file)
+        del defaults['recursive']
+    if 'ref' in defaults:
+        _validate_type_str(defaults['ref'], 'ref', yaml_file)
+        if not _valid_ref_type(defaults['ref']):
+            error = fmt.invalid_ref_error(defaults['ref'], yaml_file)
+            raise ClowderError(error)
+        del defaults['ref']
+    if 'remote' in defaults:
+        _validate_type_str(defaults['remote'], 'remote', yaml_file)
+        del defaults['remote']
+    if 'source' in defaults:
+        _validate_type_str(defaults['source'], 'source', yaml_file)
+        del defaults['source']
+    if 'depth' in defaults:
+        _validate_type_depth(defaults['depth'], yaml_file)
+        del defaults['depth']
+    if defaults:
+        error = fmt.invalid_entries_error('defaults', defaults, yaml_file)
+        raise ClowderError(error)
+
+
+def _validate_yaml_defaults(defaults, yaml_file):
+    """Validate defaults in clowder loaded from yaml file"""
+    _validate_type_dict(defaults, 'defaults', yaml_file)
+    if not defaults:
+        error = fmt.invalid_entries_error('defaults', defaults, yaml_file)
+        raise ClowderError(error)
+
+    if 'ref' not in defaults:
+        error = fmt.missing_entry_error('ref', 'defaults', yaml_file)
+        raise ClowderError(error)
+    _validate_type_str(defaults['ref'], 'ref', yaml_file)
+    if not _valid_ref_type(defaults['ref']):
+        error = fmt.invalid_ref_error(defaults['ref'], yaml_file)
+        raise ClowderError(error)
+    del defaults['ref']
+
+    if 'remote' not in defaults:
+        error = fmt.missing_entry_error('remote', 'defaults', yaml_file)
+        raise ClowderError(error)
+    _validate_type_str(defaults['remote'], 'remote', yaml_file)
+    del defaults['remote']
+
+    if 'source' not in defaults:
+        error = fmt.missing_entry_error('source', 'defaults', yaml_file)
+        raise ClowderError(error)
+    _validate_type_str(defaults['source'], 'source', yaml_file)
+    del defaults['source']
+
+    _validate_yaml_defaults_optional(defaults, yaml_file)
+
+    if defaults:
+        error = fmt.invalid_entries_error('defaults', defaults, yaml_file)
+        raise ClowderError(error)
+
+
+def _validate_yaml_defaults_optional(defaults, yaml_file):
+    """Validate defaults optional args in clowder loaded from yaml file"""
+    if 'depth' in defaults:
+        _validate_type_depth(defaults['depth'], yaml_file)
+        del defaults['depth']
+
+    if 'recursive' in defaults:
+        _validate_type_bool(defaults['recursive'], 'recursive', yaml_file)
+        del defaults['recursive']
+
+
+def _validate_yaml_fork(fork, yaml_file):
+    """Validate fork in clowder loaded from yaml file"""
+    _validate_type_dict(fork, 'fork', yaml_file)
+    if not fork:
+        error = fmt.invalid_entries_error('fork', fork, yaml_file)
+        raise ClowderError(error)
+
+    if 'name' not in fork:
+        error = fmt.missing_entry_error('name', 'fork', yaml_file)
+        raise ClowderError(error)
+    _validate_type_str(fork['name'], 'name', yaml_file)
+    del fork['name']
+
+    if 'remote' not in fork:
+        error = fmt.missing_entry_error('remote', 'fork', yaml_file)
+        raise ClowderError(error)
+    _validate_type_str(fork['remote'], 'remote', yaml_file)
+    del fork['remote']
+
+    if fork:
+        error = fmt.invalid_entries_error('fork', fork, yaml_file)
+        raise ClowderError(error)
+
+
+def _validate_yaml_import_groups(groups, yaml_file):
+    """Validate groups in clowder loaded from yaml file with import"""
+    _validate_type_list(groups, 'groups', yaml_file)
+    if not groups:
+        error = fmt.invalid_entries_error('groups', groups, yaml_file)
+        raise ClowderError(error)
+
+    for group in groups:
+        _validate_yaml_import_group(group, yaml_file)
+
+
+def _validate_yaml_groups(groups, yaml_file):
+    """Validate groups in clowder loaded from yaml file"""
+    _validate_type_list(groups, 'groups', yaml_file)
+    if not groups:
+        error = fmt.invalid_entries_error('groups', groups, yaml_file)
+        raise ClowderError(error)
+
+    for group in groups:
+        _validate_yaml_group(group, yaml_file)
+
+
+def _validate_yaml_import_project(project, yaml_file):
+    """Validate project in clowder loaded from yaml file with import"""
+    _validate_type_dict(project, 'project', yaml_file)
+    if not project:
+        error = fmt.invalid_entries_error('project', project, yaml_file)
+        raise ClowderError(error)
+
+    if 'name' not in project:
+        error = fmt.missing_entry_error('name', 'project', yaml_file)
+        raise ClowderError(error)
+    _validate_type_str(project['name'], 'name', yaml_file)
+    del project['name']
+
+    if not project:
+        error = fmt.invalid_entries_error('project', project, yaml_file)
+        raise ClowderError(error)
+
+    if 'path' in project:
+        _validate_type_str(project['path'], 'path', yaml_file)
+        del project['path']
+
+    _validate_yaml_project_optional(project, yaml_file)
+
+    if project:
+        error = fmt.invalid_entries_error('project', project, yaml_file)
+        raise ClowderError(error)
+
+
+def _validate_yaml_import_group(group, yaml_file):
+    """Validate group in clowder loaded from yaml file with import"""
+    _validate_type_dict(group, 'group', yaml_file)
+    if not group:
+        error = fmt.invalid_entries_error('group', group, yaml_file)
+        raise ClowderError(error)
+
+    if 'name' not in group:
+        error = fmt.missing_entry_error('name', 'group', yaml_file)
+        raise ClowderError(error)
+    _validate_type_str(group['name'], 'name', yaml_file)
+    del group['name']
+
+    if not group:
+        error = fmt.invalid_entries_error('group', group, yaml_file)
+        raise ClowderError(error)
+
+    if 'projects' in group:
+        _validate_yaml_projects(group['projects'], yaml_file, is_import=True)
+        del group['projects']
+
+    if 'recursive' in group:
+        _validate_type_bool(group['recursive'], 'recursive', yaml_file)
+        del group['recursive']
+
+    if 'ref' in group:
+        _validate_type_str(group['ref'], 'ref', yaml_file)
+        if not _valid_ref_type(group['ref']):
+            error = fmt.invalid_ref_error(group['ref'], yaml_file)
+            raise ClowderError(error)
+        del group['ref']
+
+    if 'remote' in group:
+        _validate_type_str(group['remote'], 'remote', yaml_file)
+        del group['remote']
+
+    if 'source' in group:
+        _validate_type_str(group['source'], 'source', yaml_file)
+        del group['source']
+
+    if 'depth' in group:
+        _validate_type_depth(group['depth'], yaml_file)
+        del group['depth']
+
+    if group:
+        error = fmt.invalid_entries_error('group', group, yaml_file)
+        raise ClowderError(error)
+
+
+def _validate_yaml_group(group, yaml_file):
+    """Validate group in clowder loaded from yaml file"""
+    _validate_type_dict(group, 'group', yaml_file)
+    if not group:
+        error = fmt.invalid_entries_error('group', group, yaml_file)
+        raise ClowderError(error)
+
+    if 'name' not in group:
+        error = fmt.missing_entry_error('name', 'group', yaml_file)
+        raise ClowderError(error)
+    _validate_type_str(group['name'], 'name', yaml_file)
+    del group['name']
+
+    if 'projects' not in group:
+        error = fmt.missing_entry_error('projects', 'group', yaml_file)
+        raise ClowderError(error)
+    _validate_yaml_projects(group['projects'], yaml_file, is_import=False)
+    del group['projects']
+
+    if 'recursive' in group:
+        _validate_type_bool(group['recursive'], 'recursive', yaml_file)
+        del group['recursive']
+
+    if 'ref' in group:
+        _validate_type_str(group['ref'], 'ref', yaml_file)
+        if not _valid_ref_type(group['ref']):
+            error = fmt.invalid_ref_error(group['ref'], yaml_file)
+            raise ClowderError(error)
+        del group['ref']
+
+    if 'remote' in group:
+        _validate_type_str(group['remote'], 'remote', yaml_file)
+        del group['remote']
+
+    if 'source' in group:
+        _validate_type_str(group['source'], 'source', yaml_file)
+        del group['source']
+
+    if 'depth' in group:
+        _validate_type_depth(group['depth'], yaml_file)
+        del group['depth']
+
+    if group:
+        error = fmt.invalid_entries_error('group', group, yaml_file)
+        raise ClowderError(error)
+
+
+def _validate_yaml_project(project, yaml_file):
+    """Validate project in clowder loaded from yaml file"""
+    _validate_type_dict(project, 'project', yaml_file)
+    if not project:
+        error = fmt.invalid_entries_error('project', project, yaml_file)
+        raise ClowderError(error)
+
+    if 'name' not in project:
+        error = fmt.missing_entry_error('name', 'project', yaml_file)
+        raise ClowderError(error)
+    _validate_type_str(project['name'], 'name', yaml_file)
+    del project['name']
+
+    if 'path' not in project:
+        error = fmt.missing_entry_error('path', 'project', yaml_file)
+        raise ClowderError(error)
+    _validate_type_str(project['path'], 'path', yaml_file)
+    del project['path']
+
+    _validate_yaml_project_optional(project, yaml_file)
+
+    if project:
+        error = fmt.invalid_entries_error('project', project, yaml_file)
+        raise ClowderError(error)
+
+
+def _validate_yaml_project_optional(project, yaml_file):
+    """Validate optional args in project in clowder loaded from yaml file"""
+    if 'remote' in project:
+        _validate_type_str(project['remote'], 'remote', yaml_file)
+        del project['remote']
+
+    if 'recursive' in project:
+        _validate_type_bool(project['recursive'], 'recursive', yaml_file)
+        del project['recursive']
+
+    if 'ref' in project:
+        _validate_type_str(project['ref'], 'ref', yaml_file)
+        if not _valid_ref_type(project['ref']):
+            error = fmt.invalid_ref_error(project['ref'], yaml_file)
+            raise ClowderError(error)
+        del project['ref']
+
+    if 'source' in project:
+        _validate_type_str(project['source'], 'source', yaml_file)
+        del project['source']
+
+    if 'depth' in project:
+        _validate_type_depth(project['depth'], yaml_file)
+        del project['depth']
+
+    if 'fork' in project:
+        fork = project['fork']
+        _validate_yaml_fork(fork, yaml_file)
+        del project['fork']
+
+
+def _validate_yaml_projects(projects, yaml_file, is_import):
+    """Validate projects in clowder loaded from yaml file"""
+    _validate_type_list(projects, 'projects', yaml_file)
+    if not projects:
+        error = fmt.invalid_entries_error('projects', projects, yaml_file)
+        raise ClowderError(error)
+
+    for project in projects:
+        if is_import:
+            _validate_yaml_import_project(project, yaml_file)
+        else:
+            _validate_yaml_project(project, yaml_file)
+
+
+def _validate_yaml_sources(sources, yaml_file):
+    """Validate sources in clowder loaded from yaml file"""
+    _validate_type_list(sources, 'sources', yaml_file)
+    if not sources:
+        error = fmt.invalid_entries_error('sources', sources, yaml_file)
+        raise ClowderError(error)
+
+    for source in sources:
+        _validate_type_dict(source, 'source', yaml_file)
+        if not source:
+            error = fmt.invalid_entries_error('source', source, yaml_file)
+            raise ClowderError(error)
+
+        if 'name' not in source:
+            error = fmt.missing_entry_error('name', 'source', yaml_file)
+            raise ClowderError(error)
+        _validate_type_str(source['name'], 'name', yaml_file)
+        del source['name']
+
+        if 'url' not in source:
+            error = fmt.missing_entry_error('url', 'source', yaml_file)
+            raise ClowderError(error)
+        _validate_type_str(source['url'], 'url', yaml_file)
+        del source['url']
+
+        if source:
+            error = fmt.invalid_entries_error('source', source, yaml_file)
+            raise ClowderError(error)
